fix(todos): return 403 for unauthorized task deletion

+++ app/services/task_service.py
@@ -161,19 +161,22 @@ class TaskService:
         user_id: int,
         task_id: int,
     ):
+
         try:
-            task = await self.task_repo.get_by_id_and_owner_id(
-                task_id,
-                user_id,
-            )
+            task = await self.task_repo.get_by_id(task_id)
 
             if not task:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=TaskMessages.TASK_NOT_FOUND,
-                )
+                raise TodoNotFoundException()
+
+            if task.owner_id != user_id:
+                raise PermissionDeniedException()
 
             await self.task_repo.delete_task(task)
+        except (
+            TodoNotFoundException,
+            PermissionDeniedException,
+        ):
+            raise
 
         except HTTPException:
             raise
